fix: apply Sprint Mission 5 code review feedback
