+++ react-project/src/pages/ItemList.module.css
@@ -1,4 +1,4 @@
-.itemListPage {
+.itemList {
   padding: 0 20px;
 }
 .container {
